Pipeline gathering subcase results
Makes some tests run 10x faster.

Subcases are changed such that the test fixture is initialized
and finalized once for all subcases. This means that subcases
that only add eventual async expectations don't block subsequent
subcases from starting. All eventual expectations from the
subcases are resolved at the end of the testcase.

This results in a bit worse device lost granularity as it
may not be possible to tell exactly where a device lost error
occured. Also, unexpected validation and OOM errors are caught
at the whole testcase level instead of the subcase level. This is
probably fine since subcases are intended to be like a "for-loop"
so exact reporting is probably not crucial.

Test finalization and releasing the device back to the
device pool occurs in the same step. This is necessary since we
shouldn't start reusing the device until eventual expectations
which operate on the same device are complete. This means a single
testcase shouldn't have complex interactions with the device pool.
To maintain this invariant, we require selecting a device with
optional formats/features to occur in the `before` stage before
any of the subcases run. Then, all subcases run using the same
device. Type narrowing is used to prevent a test function from
accessing these "setup-only" methods.